fix: export sanitizeJsonText from promptSecurity and import where used
- sanitizeJsonText was defined locally in storyBranchingController.ts
  but referenced (without import) in ai_model.utils.ts, causing
  TS2304 errors
- Moved it to promptSecurity.ts as a shared, exported utility
- Updated both files to import it from there
