+++ backend/src/utils/promptSecurity.ts
@@ -56,6 +56,18 @@ const normalizeInput = (input: string): string => {
     .replace(/\s+/g, " ") // Collapse whitespace
     .trim();
 };
+/**
+ * Strip markdown code fences (e.g. ```json ... ```) from raw AI text
+ * before attempting JSON.parse.
+ */
+export const sanitizeJsonText = (rawText: string): string => {
+  const trimmed = rawText.trim();
+  if (!trimmed.startsWith("```")) return trimmed;
+  return trimmed
+    .replace(/^```(?:json)?\s*/i, "")
+    .replace(/\s*```$/, "")
+    .trim();
+};
 
 export const validateAndFormatPrompt = (userPrompt: string): string => {
   if (!userPrompt || typeof userPrompt !== "string") {
